Added Heuristics.py script, with, oh SURPRISE, heuristics, that solves coloring problem approximately but fast (O(V^3) in a worst case scenario).
Added tests for heuristic algorithm. Most of them are similar to ones in tests for exact algorithm.

changed an empty list creation in Graph.py. This way seems to be faster

changed a check method is_adjacent_vertices in Graph.py for the simpler one

added method get_vertices in Graph.py for the sake of heuristic algorithm